refactor: move agent_uri to erc8004, move load_erc20_abi to erc20, remove core.exceptions shim
- Move bnbagent/utils/agent_uri.py → bnbagent/erc8004/agent_uri.py; delete utils/ package
  (AgentURIGenerator is ERC-8004-specific, only consumed by erc8004/agent.py)
- Move load_erc20_abi() from core/abi_loader.py → bnbagent/erc20/__init__.py
  (ERC-20 ABI belongs in the erc20 module; core/abi_loader.py now only contains create_web3)
- Remove core/exceptions.py deprecation shim (pre-release V1, no external consumers)
- Update all import paths; 343 tests pass
