+++ bnbagent/erc8004/agent.py
@@ -20,7 +20,7 @@ from web3 import Web3
 
 from ..constants import SCAN_API_URL
 from ..core.paymaster import Paymaster
-from ..utils.agent_uri import AgentURIGenerator
+from .agent_uri import AgentURIGenerator
 from ..wallets import WalletProvider
 from .constants import get_erc8004_config
 from .contract import ContractInterface
