--- bnbagent/utils/__init__.py
@@ -1,9 +1,0 @@
-"""Utility modules for bnbagent SDK."""
-
-from __future__ import annotations
-
-from .agent_uri import AgentURIGenerator
-
-__all__ = [
-    "AgentURIGenerator",
-]
